wip: use root ethblock-utils
