+++ ethblock-utils/Cargo.toml
@@ -5,6 +5,11 @@ edition = "2021"
 
 [dependencies]
 
+# revm
 revm = {git = "https://github.com/bluealloy/revm", default-features = false, features = ["serde"]}
+revm-database = {git = "https://github.com/bluealloy/revm", default-features = false, features = ["serde"]}
 alloy-eip2930 = { version = "0.1.0", features = ["serde"] } ## revmが依存しているAccessListのSerdeを有効にする
-serde = "1.0.217"
+
+# serialize
+bincode = "1.3.3"
+serde = {version = "1.0.217", features = ["derive"]}
