+++ sp1-turbo/ethblock/lib/Cargo.toml
@@ -10,7 +10,7 @@ revm-database = {git = "https://github.com/bluealloy/revm", default-features = f
 alloy-eip2930 = { version = "0.1.0", features = ["serde"] } ## revmが依存しているAccessListのSerdeを有効にする
 
 # cusom
-ethblock-utils = {path = "../utils"}
+ethblock-utils = {path = "../../../ethblock-utils"}
 
 # serialize
 bincode = "1.3.3"
