+++ sp1-turbo/ethblock/program/src/main.rs
@@ -12,7 +12,10 @@ use alloy_sol_types::SolType;
 use ethblock_lib::{trace_block, PublicValuesStruct};
 
 fn main() {
-    let b = trace_block();
+
+    let num_txs = sp1_zkvm::io::read::<usize>();
+
+    let b = trace_block(num_txs);
     assert!(b);
 
     let bytes = PublicValuesStruct::abi_encode(&PublicValuesStruct { result: b });
