+++ sp1-turbo/ethblock/script/Cargo.toml
@@ -8,13 +8,13 @@ default-run = "sha2"
 name = "sha2"
 path = "src/bin/main.rs"
 
-[[bin]]
-name = "evm"
-path = "src/bin/evm.rs"
+# [[bin]]
+# name = "evm"
+# path = "src/bin/evm.rs"
 
-[[bin]]
-name = "vkey"
-path = "src/bin/vkey.rs"
+# [[bin]]
+# name = "vkey"
+# path = "src/bin/vkey.rs"
 
 [dependencies]
 sp1-sdk = "4.0.0"
